ikiAI_Agents/agents/time_agent.py
from flask import Flask
from config import Config
from chroma import ChromaDB

app = Flask(__name__)
app.config.from_object(Config)
db = ChromaDB(app)

class TimeAgent:
    def __init__(self, user_id):
        self.user_id = user_id

    def monitor_user_performance(self):
        # Implementation of user performance monitoring
        pass

    def get_task_completion_percentage(self):
        # Implementation of task completion percentage retrieval
        pass

    def track_daily_task_completion(self):
        # Implementation of daily task completion tracking
        pass

    def send_goal_completion_notifications(self):
        # Implementation of goal completion notification sending
        pass

    def store_goal_in_database(self, goals):
        # Implementation of goal storing in database
        pass

    def retrieve_goals_from_database(self):
        # Implementation of goal retrieval from database
        pass

    def send_motivational_messages(self):
        # Implementation of motivational message sending
        pass
